requirements.txt
# Core dependencies
torch==2.2.2
pyro-ppl==1.9.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.2
pyyaml==6.0.1

# Jupyter and notebook support
jupyter==1.0.0
notebook==7.1.3
ipykernel==6.29.4

# Visualization
matplotlib==3.8.4
seaborn==0.13.2

# Development and testing
pytest==8.1.1
black==24.3.0
flake8==7.0.0
mypy==1.9.0

# SHAP for interpretability
shap==0.45.0